refactored footer && navigation components, added lang changing to mobile ver
